requirements.txt
requests==2.32.3
urllib3==2.2.3
python-dotenv==1.0.1